Request for third button in modal window #1728
