+++ stencil-workspace/src/components.d.ts
@@ -9,7 +9,7 @@ import { ModusAutocompleteOption } from "./components/modus-autocomplete/modus-a
 import { Crumb } from "./components/modus-breadcrumb/modus-breadcrumb";
 import { ModusDataTableCellLink, ModusDataTableDisplayOptions, ModusDataTableRowAction, ModusDataTableRowActionClickEvent, ModusDataTableSelectionOptions, ModusDataTableSortEvent, ModusDataTableSortOptions, TCell, TColumn, TRow } from "./components/modus-data-table/modus-data-table.models";
 import { ModusDateInputEventDetails, ModusDateInputType } from "./components/modus-date-input/utils/modus-date-input.models";
-import { ModusModalButtons } from "./components/modus-modal/modus-modal";
+import { ModusModalButtons } from "./components/modus-modal/modus-button";
 import { ModusNavbarApp } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
 import { ModusNavbarButton, ModusNavbarLogoOptions, ModusNavbarProfileMenuLink, ModusNavbarTooltip, ModusProfileMenuOptions } from "./components/modus-navbar/modus-navbar.models";
 import { ModusNavbarApp as ModusNavbarApp1 } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
@@ -24,7 +24,7 @@ export { ModusAutocompleteOption } from "./components/modus-autocomplete/modus-a
 export { Crumb } from "./components/modus-breadcrumb/modus-breadcrumb";
 export { ModusDataTableCellLink, ModusDataTableDisplayOptions, ModusDataTableRowAction, ModusDataTableRowActionClickEvent, ModusDataTableSelectionOptions, ModusDataTableSortEvent, ModusDataTableSortOptions, TCell, TColumn, TRow } from "./components/modus-data-table/modus-data-table.models";
 export { ModusDateInputEventDetails, ModusDateInputType } from "./components/modus-date-input/utils/modus-date-input.models";
-export { ModusModalButtons } from "./components/modus-modal/modus-modal";
+export { ModusModalButtons } from "./components/modus-modal/modus-button";
 export { ModusNavbarApp } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
 export { ModusNavbarButton, ModusNavbarLogoOptions, ModusNavbarProfileMenuLink, ModusNavbarTooltip, ModusProfileMenuOptions } from "./components/modus-navbar/modus-navbar.models";
 export { ModusNavbarApp as ModusNavbarApp1 } from "./components/modus-navbar/apps-menu/modus-navbar-apps-menu";
@@ -571,7 +571,7 @@ export namespace Components {
          */
         "backdrop": 'default' | 'static';
         /**
-          * The modal button to render.
+          * The modal's footer buttons.
          */
         "buttons": ModusModalButtons;
         /**
@@ -2436,7 +2436,7 @@ declare namespace LocalJSX {
          */
         "backdrop"?: 'default' | 'static';
         /**
-          * The modal button to render.
+          * The modal's footer buttons.
          */
         "buttons"?: ModusModalButtons;
         /**
