+++ stencil-workspace/src/components/modus-modal/modus-button.tsx
@@ -1,0 +1,14 @@
+
+export interface ModusModalButtons {
+  primary?: ModusModalButton;
+  secondary?: ModusModalButton;
+  outline?: ModusModalButton;
+}
+interface ModusModalButton {
+  /** (optional) The modal's  button text. */
+  text?: string;
+  /** (optional) The modal's  button aria-label. */
+  ariaLabel?: string;
+  /** (optional) Disable  button. */
+  disabled?: boolean;
+}
