+++ stencil-workspace/src/components/modus-modal/modus-modal.e2e.ts
@@ -1,5 +1,6 @@
 import { newE2EPage } from '@stencil/core/testing';
-import { ModusModalButtons } from './modus-modal';
+import {ModusModalButtons} from "./modus-button";
+
 
 describe('modus-modal', () => {
   it('renders', async () => {
